refactor(sleep): split SleepScreen UI into sibling files
Re-cut of the extraction in #734 (by @tigercraft4) against current main, which had
moved on ~388 commits (SleepScreen.kt rewritten +608/-103 since that branch point).
Pure UI code-motion, no behaviour change: moves three clusters out of SleepScreen.kt
into siblings, keeping each comment with its code:

- SleepNightNavUi.kt — sleep window row + night navigation / edit / delete controls
- SleepMetricCardsUi.kt — metric grid, debt ledger, stage comparison, trend cards,
  shared Sleep chart helpers
- SleepMetricDetailSheet.kt — the metric detail bottom sheet

Fixes over the original: adds the missing Polish translation for the two delete-dialog
body strings (values-pl is gated at zero), and trims/dedupes imports in the touched
files. Cross-file helpers widened private->internal; the generic helper names that
collide across screens (Hairline/ChartCard/ChartFooter/LegendDot) are renamed to
Sleep-prefixed internals so they do not clash with other screens private copies.

No sleep calculations, storage, BLE, or screen wiring changed. compileFullDebugKotlin
+ compileFullDebugUnitTestKotlin + processFullDebugResources + i18n --ci all green.
